Update of Material/Zaid for correct merge with MX cards
       Big update since the zaid system for mx cards
       was a mess.
       -- All in size_t
       -- MXcards now a map
       Todo : still to look if we have a problem with two different
       types of zaid in one material e.g. 1001.70c and 1001.80c

+++ System/monteInc/Material.h
@@ -37,7 +37,7 @@ namespace MonteCarlo
 */
 
 class Material 
-{
+{ 
  private:
   
   int Mnum;                        ///< Material Number (Necessary)  
@@ -52,8 +52,8 @@ class Material
   int getExtraType(std::string&,std::string&);
   void calcAtomicDensity();
 
-  size_t getZaidIndex(const int,const int,const char) const;
-  static void writeZaid(std::ostream&,const double,const int);
+  size_t getZaidIndex(const size_t,const size_t,const char) const;
+  static void writeZaid(std::ostream&,const double,const size_t);
 
  public:
   
@@ -97,7 +97,7 @@ class Material
   int setMaterial(const int,const std::string&,const std::string&,
 		  const std::string&);
 
-  void setMXitem(const int,const int,const char,const std::string&,
+  void setMXitem(const size_t,const size_t,const char,const std::string&,
 		 const std::string&);
   /// Get atomic density
   double getAtomDensity() const { return atomDensity; }
@@ -105,14 +105,14 @@ class Material
   double getMeanA() const;
   void setENDF7();
   void setDensity(const double);
-  bool hasZaid(const int,const int,const char) const;
+  bool hasZaid(const size_t,const size_t,const char) const;
 
   /// remove mt cards
   void removeSQW() { SQW.clear(); }
   void removeMX(const std::string&);
   void removeLib(const std::string&);
     
-  void changeLibrary(const int,const char);
+  void changeLibrary(const size_t,const char);
 
   void listComponent() const;
   void print() const;
